--- viewer/data/proteusPy-0.97.11.whl
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:79312a57695493392e2cbe72add7848d552babb099f400f63705873c44b4dd1b
-size 373178
