--- viewer/data/proteuspy.tar.gz
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:74fb286ff28471deb695108644d68817bd54ce0701bedcbd87a6062af9576fbe
-size 371366
