+++ viewer/proteuspy-0.97.12.dev0.macosx-11.0-arm64.tar.gz
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:9a4028c4b19b5d17ec7f81ca9f9db7abb01756690b170f1c4ba2473033302671
+size 515546
